requirements.txt

fastapi==0.115.11

PyYAML==6.0.2

uvicorn==0.34.0
